Reorganize redux file structure
