+++ client/src/reducers.js
@@ -1,5 +1,5 @@
 import { combineReducers } from 'redux';
-import projectListReducer from './projectListReducer';
+import projectListReducer from './main/reducers/projectListReducer';
 
 export default combineReducers({
   projectList: projectListReducer,
